fixing issues

+++ web_scraping/functions/bucket_helpers.py
@@ -1,22 +1,27 @@
-from concurrent.futures import ThreadPoolExecutor
 import requests
 import pandas as pd
 import json
 from pathlib import Path
 import sys
+import io
 
 # Set base path for helper functions
 PROJECT_DIR = Path(__file__).parent.parent.parent
 sys.path.append(str(PROJECT_DIR))
 
 # Import bucket connection
-from utils.connections import bucket, bucket_name
+from utils.connections import bucket
 
 
 def load_links_to_gcs(dct):
     content = json.dumps(dct, indent=4)
     destination = "web-scraping/scraped_links.json"
     blob = bucket.blob(destination)
+
+    # now sure if it works to get the latest version of file
+    if blob.exists():
+        blob.delete()
+
     blob.upload_from_string(content, content_type="application/json")
 
 
@@ -34,50 +39,46 @@ def load_img_to_gcs(link, img_name):
     blob.upload_from_string(content, content_type="image/jpeg")
 
 
-def load_file_to_gcs(data, file_name, csv=True):
-    destination = f"web-scraping/raw/{file_name}"
+def load_file_to_gcs(data, path, file_name, csv=True):
+    destination = f"{path}/{file_name}"
     blob = bucket.blob(destination)
 
+    if blob.exists():
+        blob.delete()
+
     if csv:
-        df = pd.DataFrame(data)
-        blob.upload_from_string(df.to_csv(index=False), content_type="text/csv")
+        blob.upload_from_string(data.to_csv(index=False), content_type="text/csv")
     else:
         content = json.dumps(data, indent=4)
         blob.upload_from_string(content, content_type="application/json")
 
 
-def open_file_from_gcs(file_name, csv=True):
+def open_file_from_gcs(path, file_name, csv=True):
+    blob = bucket.blob(f"{path}/{file_name}")
+
     if csv:
-        path = f"gs://{bucket_name}/{web_scraping_path}raw/{file_name}"
-        df = pd.read_csv(path)
+        data = blob.download_as_bytes()
+        df = pd.read_csv(io.BytesIO(data))
         return df
     else:
-        blob = bucket.blob(f"{web_scraping_path}raw/{file_name}")
-        data = json.loads(blob.download_as_string(client=None))
+        data = json.loads(blob.download_as_string())
         return data
 
 
-def move_image_to_gcs():
+def move_image_gcs():
     categories = ["boots", "balls"]
     num = 0
 
-    # helper func to move and delete single blob
-    def move_and_delete_blob(blob):
-        blob_name = "/".join(blob.name.split("/")[3:])
-        destination = f"product-images/{blob_name}"
-        bucket.copy_blob(blob, bucket, destination)
-        bucket.delete_blob(blob.name)
-
-    # retrieve all files from folder
     for category in categories:
-        prefix = f"{web_scraping_path}img/{category}/"
-        blobs = [blob for blob in bucket.list_blobs(prefix=prefix) if blob.name.endswith(".jpg")]
-        num += len(blobs)
+        prefix = f"web-scraping/img/{category}"
 
-        # parallel execution of moving and deleting blobs
-        with ThreadPoolExecutor(max_workers=10) as exec:
-            exec.map(move_and_delete_blob, blobs)
-
+        for blob in bucket.list_blobs(prefix=prefix):
+            if blob.name.endswith(".jpg"):
+                blob_name = "/".join(blob.name.split("/")[3:])
+                destination = f"product-images/{blob_name}"
+                bucket.copy_blob(blob, bucket, destination)
+                bucket.delete_blob(blob.name)
+                num += 1
     return num
 
 
@@ -87,8 +88,8 @@ def delete_blobs_from_gcs():
 
     # retrieve all files from folder
     for category in categories:
-        prefix = f"{web_scraping_path}img/{category}/"
-        blobs = [blob for blob in bucket.list_blobs(prefix=prefix) if blob.name.endswith(".jpg")]
+        prefix = f"web-scraping/img/{category}"
 
-        with ThreadPoolExecutor(max_workers=10) as exec:
-            exec.map(lambda blob: bucket.delete_blob(blob.name), blobs)
+        for blob in bucket.list_blobs(prefix=prefix):
+            if blob.name.endswith(".jpg"):
+                bucket.delete_blob(blob.name)
